feat: rename email component file
